refactor: modularize demo-oauth2 for better code organization
- Split main.rs into focused modules:
  - state.rs: App state and auth-related types
  - handlers.rs: Route handlers with their templates
  - server.rs: Server configuration and spawning

- Improve code organization:
  - Move template structs closer to their handlers
  - Add pub(crate) visibility for internal APIs
  - Clean up imports and remove unused ones

This refactoring improves code maintainability by separating
concerns into focused modules while keeping related code together.
