+++ demo-oauth2/src/server.rs
@@ -1,0 +1,43 @@
+use axum::Router;
+use axum_server::tls_rustls::RustlsConfig;
+use std::{net::SocketAddr, path::PathBuf};
+use tokio::task::JoinHandle;
+
+#[derive(Clone, Copy)]
+pub(crate) struct Ports {
+    pub(crate) http: u16,
+    pub(crate) https: u16,
+}
+
+pub(crate) fn spawn_http_server(port: u16, app: Router) -> JoinHandle<()> {
+    tokio::spawn(async move {
+        let addr = SocketAddr::from(([0, 0, 0, 0], port));
+        tracing::debug!("HTTP server listening on {}:{}", addr, port);
+        axum_server::bind(addr)
+            .serve(app.into_make_service())
+            .await
+            .unwrap();
+    })
+}
+
+pub(crate) async fn spawn_https_server(port: u16, app: Router) -> JoinHandle<()> {
+    let config = RustlsConfig::from_pem_file(
+        PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+            .join("self_signed_certs")
+            .join("cert.pem"),
+        PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+            .join("self_signed_certs")
+            .join("key.pem"),
+    )
+    .await
+    .unwrap();
+
+    let addr = SocketAddr::from(([0, 0, 0, 0], port));
+    tracing::debug!("HTTPS server listening on {}:{}", addr, port);
+    tokio::spawn(async move {
+        axum_server::bind_rustls(addr, config)
+            .serve(app.into_make_service())
+            .await
+            .unwrap();
+    })
+}
